requirements.txt
fastapi
uvicorn
pydantic
openai
chromadb
duckduckgo-search
streamlit
python-dotenv
